Refactor: 生成APIとテキストのセーブを一つのAPIにまとめる
